Fix C++ drivers: add missing STL includes and parser fixes
Adds missing `#include` directives that caused compile errors when user
code didn't transitively include them (`<algorithm>`, `<queue>`,
`<vector>`, `<unordered_map>`, etc.). Fixes a uint32 parsing overflow in
count-bits (Java/C++) that rejected values > INT_MAX. Corrects
equation-solver 2D-array parser that was skipping the first inner `[`.
Renames `int2DVector` to `intVector2D` in height-queue and zero-matrix
for consistency.

Verified via Judge0: all affected problems now pass 100% across Java,
Python, JavaScript, and C++.

+++ common/src/main/resources/drivers/count-bits/cpp.cpp
@@ -4,9 +4,9 @@ using namespace std;
 {USER_CODE}
 
 int main() {
-    int n;
+    unsigned int n;
     cin >> n;
-    int result = Solution().hammingWeight(n);
+    int result = Solution().hammingWeight((int)n);
     cout << result << endl;
     return 0;
 }
